feature/places_sezam

+++ dao/src/main/java/greencity/repository/BreakTimeRepository.java
@@ -3,5 +3,5 @@ package greencity.repository;
 import greencity.entity.BreakTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BreakTimeRepository extends JpaRepository<BreakTime, Integer> {
+public interface BreakTimeRepository extends JpaRepository<BreakTime, Long> {
 }
