+++ dao/src/main/java/greencity/repository/DiscountValueRepository.java
@@ -3,5 +3,5 @@ package greencity.repository;
 import greencity.entity.DiscountValue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscountValueRepository extends JpaRepository<DiscountValue, Integer> {
+public interface DiscountValueRepository extends JpaRepository<DiscountValue, Long> {
 }
